requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2024.2.2
crispy-bootstrap4==2024.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.9
django-crispy-forms==2.1
django-import-export==3.3.8
et-xmlfile==1.1.0
greenlet==3.0.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.3.0
python-dotenv==1.0.1
python-telegram-bot==13.11
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
tablib==3.5.0
tornado==6.4
typing_extensions==4.11.0
tzlocal==5.2
xlrd==2.0.1
xlwt==1.3.0
